requirements.txt
# requirements.txt
# Python Dependencies for LegacyParts AI Backend

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep Learning & Computer Vision
torch==2.1.0
torchvision==0.16.0
onnx==1.15.0
onnx-tf==1.10.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# API & Data Handling
pydantic==2.5.0
requests==2.31.0
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Science & Analytics
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# Kaggle Integration
kaggle==1.5.16

# Monitoring & Logging
sentry-sdk==1.38.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Deployment
gunicorn==21.2.0
docker==6.1.3

# Optional: If using TensorFlow.js conversion
# tensorflow==2.15.0
# tensorflowjs==4.14.0
